+++ src/eml2/Activity.h
@@ -115,7 +115,7 @@ namespace EML2_NS
 		 *
 		 * @returns	The parameter count.
 		 */
-		DLL_IMPORT_OR_EXPORT virtual unsigned int getParameterCount() const = 0;
+		DLL_IMPORT_OR_EXPORT virtual uint64_t getParameterCount() const = 0;
 
 		/**
 		 * Gets the count of all the parameters which have the same title.
@@ -124,7 +124,7 @@ namespace EML2_NS
 		 *
 		 * @returns	The parameter count.
 		 */
-		DLL_IMPORT_OR_EXPORT virtual unsigned int getParameterCount(const std::string & paramTitle) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual uint64_t getParameterCount(const std::string & paramTitle) const = 0;
 
 		/**
 		 * Gets the title of a given parameter
@@ -236,7 +236,7 @@ namespace EML2_NS
 		 *
 		 * @returns	A vector of the value of all the @p paramTitle integer quantity parameters.
 		 */
-		DLL_IMPORT_OR_EXPORT virtual std::vector<int64_t> getIntegerQuantityParameterValue(const std::string & paramTitle) const = 0;
+		DLL_IMPORT_OR_EXPORT virtual std::vector<int32_t> getIntegerQuantityParameterValue(const std::string & paramTitle) const = 0;
 
 		/**
 		 * Gets the integer quantity value of a given parameter.
