+++ src/eml2_3/ActivityTemplate.cpp
@@ -91,7 +91,7 @@ bool ActivityTemplate::isAnExistingParameter(const std::string & paramTitle) con
 	return false;
 }
 
-unsigned int ActivityTemplate::getParameterCount() const
+uint64_t ActivityTemplate::getParameterCount() const
 {
 	return static_cast<eml23__ActivityTemplate*>(gsoapProxy2_3)->Parameter.size();
 }
